refactor: Move method from lotto to user for Separation of Responsibilities
